vesc/application: send status1 message at twice the configured rate
The status 1 message contains RPM information which is very useful
for Motion to track the wheels. Sending it faster allows motion
to monitor the RPM more precisely during a ball launch and
potentially detect ball jams while the wheels are still spinning.

Clean up some of the status message level code to consolidate
the if statements to make the overall code easier to read.

Remove some of the dual motor code to make the remaining code
more readable. Out hardware does not support dual motors and
we have no plans to do so in the future.